# docs/openapi/marketplace.yaml
openapi: 3.1.0
info:
  title: BMPL Marketplace API
  version: "2.0.0"
  description: >
    Belize Marketplace & Logistics — Phase 2 marketplace endpoints (categories,
    vendors, products, images, variants, inventory, public catalog/search).
    All paths are under the `/api` prefix. Money is integer minor units (BZD cents).
servers:
  - url: https://bmplapi-production.up.railway.app/api
    description: Production
  - url: http://localhost:4000/api
    description: Local
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
...
  responses:
...
    Unauthorized: { description: Authentication required }
...
    NotFound: { description: Not found or not owned }
...
paths:
...
  /driver/jobs/{id}:
    get: { tags: [Dispatch], summary: "Own job detail (404 for another driver's job)", parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [] }], responses: { "200": { description: OK }, "404": { $ref: "#/components/responses/NotFound" } } }
...
  /saved:
    get: { tags: [Engagement], summary: "The caller's saved products (newest first, resolved cards)", security: [{ cookieAuth: [] }], responses: { "200": { description: "{ items: [{ productId, savedAt, available, product }] }" }, "401": { $ref: "#/components/responses/Unauthorized" } }, description: "Requires CUSTOMER role; unavailable saved products return available:false, product:null (row preserved)" }
  /saved/ids:
    get: { tags: [Engagement], summary: "Lightweight set of saved product ids (heart-toggle state)", security: [{ cookieAuth: [] }], responses: { "200": { description: "{ productIds: string[] }" } } }
...
  /saved/{productId}:
    post: { tags: [Engagement], summary: "Save a product (idempotent)", parameters: [ { name: productId, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: "{ saved: true }" }, "404": { description: "Product not viewable" } } }
    delete: { tags: [Engagement], summary: "Remove a saved product (idempotent)", parameters: [ { name: productId, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "200": { description: "{ saved: false }" } } }
  /recently-viewed:
    get: { tags: [Engagement], summary: "The caller's recently-viewed history (newest first)", parameters: [ { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 50 } } ], security: [{ cookieAuth: [] }], responses: { "200": { description: "{ items: [{ productId, viewedAt, available, product }] }" } }, description: "Private to the owner; capped to the newest 50" }
    delete: { tags: [Engagement], summary: "Clear the caller's recently-viewed history", security: [{ cookieAuth: [], csrfToken: [] }], responses: { "200": { description: "{ ok: true }" } } }
  /recently-viewed/{productId}:
    post: { tags: [Engagement], summary: "Record a product view (upsert; prunes to newest 50)", parameters: [ { name: productId, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: "{ ok: true }" }, "404": { description: "Product not viewable" } } }
  # ---- Discovery & Recommendations (Phase 4 · M21) ----
...
  /jobs:
    get: { tags: [Jobs], summary: "Public job search", parameters: [ { name: q, in: query, schema: { type: string } }, { name: category, in: query, schema: { type: string } }, { name: district, in: query, schema: { type: string } }, { name: employmentType, in: query, schema: { type: string } }, { name: workArrangement, in: query, schema: { type: string } }, { name: remote, in: query, schema: { type: boolean } }, { name: salaryMin, in: query, schema: { type: number } }, { name: sort, in: query, schema: { type: string, enum: [relevance, newest, deadline, salary_asc, salary_desc] } }, { name: page, in: query, schema: { type: integer } } ], responses: { "200": { description: "{ total, page, pageSize, items }" } }, description: "PUBLISHED jobs of APPROVED employers only; excludes expired/closed/draft" }
...
  /jobs/companies/{slug}:
    get: { tags: [Jobs], summary: "Public company page + open jobs", parameters: [ { name: slug, in: path, required: true, schema: { type: string } } ], responses: { "200": { description: OK }, "404": { $ref: "#/components/responses/NotFound" } } }
  /jobs/{slug}:
    get: { tags: [Jobs], summary: "Public job detail (+ company + related)", parameters: [ { name: slug, in: path, required: true, schema: { type: string } } ], responses: { "200": { description: OK }, "404": { $ref: "#/components/responses/NotFound" } } }
  /job-seeker/profile:
    get: { tags: [Jobs], summary: "Own job-seeker profile", security: [{ cookieAuth: [] }], responses: { "200": { description: OK } } }
    put: { tags: [Jobs], summary: "Upsert job-seeker profile (grants JOB_SEEKER)", security: [{ cookieAuth: [], csrfToken: [] }], responses: { "200": { description: OK } } }
  /job-seeker/resumes/presign:
    post: { tags: [Jobs], summary: "Presigned private PUT for a résumé (PDF/DOCX; rate-limited)", security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: "{ uploadUrl, key }" }, "400": { description: "Disallowed type" } } }
  /job-seeker/resumes:
    post: { tags: [Jobs], summary: "Confirm an uploaded résumé", security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: OK } }, description: "body { storageKey, label }" }
  /job-seeker/resumes/{id}/url:
    get: { tags: [Jobs], summary: "Signed URL for OWN résumé", parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [] }], responses: { "200": { description: "{ url }" } } }
  /job-seeker/saved/{jobId}:
    post: { tags: [Jobs], summary: "Save a job", parameters: [ { name: jobId, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: "{ saved: true }" } } }
    delete: { tags: [Jobs], summary: "Unsave a job", parameters: [ { name: jobId, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "200": { description: "{ saved: false }" } } }
  /job-seeker/applications:
    get: { tags: [Jobs], summary: "The caller's own applications", security: [{ cookieAuth: [] }], responses: { "200": { description: OK } } }
    post: { tags: [Jobs], summary: "Submit an application (rate-limited)", security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: OK }, "400": { description: "Duplicate / required question / deadline / résumé ownership" }, "404": { description: "Job not accepting" } }, description: "body { jobId, resumeId?, coverLetter?, answers:[{questionId,text?,choices?}] }" }
  /job-seeker/applications/{id}/withdraw:
    post: { tags: [Jobs], summary: "Withdraw own application", parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: OK } } }
  /job-seeker/applications/{id}/conversation:
    post: { tags: [Jobs], summary: "Open the employer↔applicant thread", parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: Conversation } } }
  /job-seeker/report/{jobId}:
    post: { tags: [Jobs], summary: "Report a job listing", parameters: [ { name: jobId, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: "{ ok: true }" } }, description: "body { reason, note? }" }
...